css/css-regions/parsing/region-fragment-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Regions Module Level 1: getComputedValue().regionFragment</title>
<link rel="help" href="https://drafts.csswg.org/css-regions/#propdef-region-fragment">
<meta name="assert" content="region-fragment computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("region-fragment", "auto");
test_computed_value("region-fragment", "break");
</script>
</body>
</html>
